Server/requirements.txt
Adafruit-Blinka==8.26.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.10
adafruit-circuitpython-pixelbuf==2.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.56.0
Adafruit-PureIO==1.1.11
blinker==1.7.0
click==8.1.7
flask==3.0.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
